issue10    separated test cases according to their types and resolved minor conflicts
