+++ actionlist/__tests__/components/actionItemBodySpec.js
@@ -4,11 +4,11 @@ import { Text, View, Button } from 'react-native';
 import { Provider } from 'react-redux';
 import { expect } from 'chai';
 import configureStore from 'redux-mock-store';
-import ActionItemBody from '../app/components/action_item_body';
+import ActionItemBody from '../../app/components/action_item_body';
 
-require('../testConfig');
+require('../../testConfig');
 
-jest.unmock('../app/components/action_item_body');
+jest.unmock('../../app/components/action_item_body');
 
 describe('Action_Item_Body', () => {
   it('correctly passes our props to Text Components in the View Component', () => {
