+++ actionlist/__tests__/components/headerSpec.js
@@ -8,22 +8,11 @@ import { mount } from 'enzyme';
 import { Provider } from 'react-redux';
 import { expect } from 'chai';
 import configureStore from 'redux-mock-store';
-import Header from '../app/components/header';
+import Header from '../../app/components/header';
 
-require('react-native-mock-render/mock');
+require('../../testConfig');
 
-jest.unmock('redux-mock-store');
-jest.unmock('../app/components/header');
-
-const jsdom = require('jsdom').jsdom;
-
-global.document = jsdom('');
-global.window = document.defaultView;
-Object.keys(document.defaultView).forEach((property) => {
-  if (typeof global[property] === 'undefined') {
-    global[property] = document.defaultView[property];
-  }
-});
+jest.unmock('../../app/components/header');
 
 describe('header', () => {
   it('renders the component tree correctly with all default component props', () => {
@@ -34,7 +23,7 @@ describe('header', () => {
     const icons = {
       /* eslint-disable global-require */
       /* rule disabled since image loading need not be global */
-      logo: require('../app/components/img/trident-large.png'),
+      logo: require('../../app/components/img/trident-large.png'),
       /* eslint-enable global-require */
     };
     const wrapper = mount(
