+++ actionlist/__tests__/reducers/actionItemsSpec.js
@@ -1,8 +1,8 @@
 import deepFreeze from 'deep-freeze';
 import expect from 'expect';
 
-import reducer from '../app/reducers/action_items';
-import { toggleDrawer } from '../app/actions/action_items';
+import reducer from '../../app/reducers/action_items';
+import { toggleDrawer } from '../../app/actions/action_items';
 
 describe('action_items', () => {
   it('should have the opposite drawerExpanded value when toggled', () => {
